feat: implement CreateAccountForm and refactor LoginForm to use new components
